Link rewrite complete
