+++ opencosmo/collection/__init__.py
@@ -1,9 +1,8 @@
 from .collection import Collection, ParticleCollection, SimulationCollection
-from .io import open_linked, open_multi_dataset_file, read_multi_dataset_file
+from .io import open_multi_dataset_file, read_multi_dataset_file
 
 __all__ = [
     "Collection",
-    "open_linked",
     "open_multi_dataset_file",
     "read_multi_dataset_file",
     "ParticleCollection",
