+++ opencosmo/link/__init__.py
@@ -1,0 +1,5 @@
+from .collection import LinkedCollection
+from .handler import LinkHandler, OomLinkHandler
+from .io import open_linked_files, open_linked_file
+
+__all__ = ['LinkedCollection', 'LinkHandler', 'OomLinkHandler', 'open_linked_files', 'open_linked_file']
